Fixed a problem where static event wasn't properly disposed of, some other stuff aswell, fixed a bug where Game didnt properly calculate games needed to win when the player disarms a shield

+++ Assets/Scripts/MainMenu/LevelManager.cs
@@ -47,7 +47,7 @@ public class LevelManager : MonoBehaviour
             float progress = Mathf.Clamp01(asyncOperation.progress / 0.9f);
 
             slider.value = progress;
-            progressText.text = progress * 100f + "%";
+            progressText.text = (int)(progress * 100) + "%";
 
             yield return null;
         }
